Make stylistic markup available to all templates
